chore: reorganize ethereum service
