+++ packages/element-lib/src/adapters/blockchain/ethereum/__tests__/blockchain.ethereum.spec.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-underscore-dangle */
 const element = require('../../../../../index');
 
 const config = require('../../../json/config.local.json');
@@ -23,11 +24,13 @@ describe('blockchain.ethereum', () => {
 
     it('can create new contracts on the fly', async () => {
       const accounts = await blockchain.web3.eth.getAccounts();
-      const instance = await blockchain.createNewContract(accounts[0]);
+      // FIXME
+      const instance = await blockchain._createNewContract(accounts[0]);
       expect(blockchain.anchorContractAddress).toBe(instance.address);
     });
   });
 
+  // FIXME
   describe('getBlockchainTime', () => {
     it('should return hash and blocknumber', async () => {
       const result = await blockchain.getBlockchainTime(0);
